:truck: Move namespace-related classes into a new subfolder
