+++ midi-script/AbletonJS.py
@@ -5,34 +5,34 @@ import time
 import Live
 from _Framework.ControlSurface import ControlSurface
 
-from .Application import Application
-from .ApplicationView import ApplicationView
-from .Browser import Browser
-from .BrowserItem import BrowserItem
-from .Chain import Chain
-from .Clip import Clip
-from .ClipSlot import ClipSlot
 from .Config import DEBUG
-from .CuePoint import CuePoint
-from .Device import Device
-from .DeviceParameter import DeviceParameter
-from .DeviceView import DeviceView
-from .DrumPad import DrumPad
-from .Envelope import Envelope
-from .Interface import Interface
-from .Internal import Internal
 from .Logging import logger
-from .LooperDevice import LooperDevice
-from .Midi import Midi
-from .MixerDevice import MixerDevice
-from .PluginDevice import PluginDevice
-from .Scene import Scene
-from .Session import Session
+from .Namespaces.Application import Application
+from .Namespaces.ApplicationView import ApplicationView
+from .Namespaces.Browser import Browser
+from .Namespaces.BrowserItem import BrowserItem
+from .Namespaces.Chain import Chain
+from .Namespaces.Clip import Clip
+from .Namespaces.ClipSlot import ClipSlot
+from .Namespaces.CuePoint import CuePoint
+from .Namespaces.Device import Device
+from .Namespaces.DeviceParameter import DeviceParameter
+from .Namespaces.DeviceView import DeviceView
+from .Namespaces.DrumPad import DrumPad
+from .Namespaces.Envelope import Envelope
+from .Namespaces.Interface import Interface
+from .Namespaces.Internal import Internal
+from .Namespaces.LooperDevice import LooperDevice
+from .Namespaces.Midi import Midi
+from .Namespaces.MixerDevice import MixerDevice
+from .Namespaces.PluginDevice import PluginDevice
+from .Namespaces.Scene import Scene
+from .Namespaces.Session import Session
+from .Namespaces.Song import Song
+from .Namespaces.SongView import SongView
+from .Namespaces.Track import Track
+from .Namespaces.TrackView import TrackView
 from .Socket import Socket
-from .Song import Song
-from .SongView import SongView
-from .Track import Track
-from .TrackView import TrackView
 from .version import version
