+++ midi-script/Namespaces/Clip.py
