+++ midi-script/Namespaces/ClipSlot.py
