+++ midi-script/Namespaces/Envelope.py
