+++ midi-script/Namespaces/Internal.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import
+from ..version import version
 from .Interface import Interface
-from .version import version
 
 
 class Internal(Interface):
