+++ midi-script/Namespaces/__init__.py
@@ -1,0 +1,1 @@
+from __future__ import absolute_import
